Convert to PyPI application.
