+++ bidskit.egg-info/entry_points.txt
@@ -1,3 +1,3 @@
 [console_scripts]
-bidskit = bidskit:main
+bidskit = bidskit.launcher:main
