+++ bidskit/launcher.py
@@ -29,13 +29,13 @@ Without Session Subdirectories:
 
 Usage
 ----
-main.py -i <DICOM Directory>[dicom] -o <BIDS Source Directory>[source] [--no-sessions] [--overwrite]
+launcher.py -i <DICOM Directory>[dicom] -o <BIDS Source Directory>[source] [--no-sessions] [--overwrite]
 
 Examples
 ----
-% main.py
-% main.py --no-sessions
-% main.py -i mydicom -o mybids --no-sessions
+% launcher.py
+% launcher.py --no-sessions
+% launcher.py -i mydicom -o mybids --no-sessions
 
 Authors
 ----
@@ -73,7 +73,7 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 """
 
-__version__ = '1.1.3'
+__version__ = '1.2'
 
 import os
 import sys
@@ -81,9 +81,10 @@ import argparse
 import subprocess
 import shutil
 import json
-import pydicom
 import numpy as np
 from glob import glob
+
+import bidskit.io as bio
 
 
 def main():
@@ -122,9 +123,9 @@ def main():
     work_dir = os.path.join(bids_root_dir, 'work', 'conversion')
 
     # Safely create the BIDS working, source and derivatives directories
-    safe_mkdir(work_dir)
-    safe_mkdir(bids_src_dir)
-    safe_mkdir(bids_deriv_dir)
+    bio.safe_mkdir(work_dir)
+    bio.safe_mkdir(bids_src_dir)
+    bio.safe_mkdir(bids_deriv_dir)
 
     print('')
     print('------------------------------------------------------------')
@@ -143,7 +144,7 @@ def main():
     # and a template will be created in the derivatives/conversion directory.
     # This template should be completed by the user and the conversion rerun.
     prot_dict_json = os.path.join(bids_deriv_dir, 'Protocol_Translator.json')
-    prot_dict = bids_load_prot_dict(prot_dict_json)
+    prot_dict = bio.bids_load_prot_dict(prot_dict_json)
 
     if prot_dict and os.path.isdir(work_dir):
         print('')
@@ -160,7 +161,7 @@ def main():
 
     # Initialize BIDS source directory contents
     if not first_pass:
-        bids_init(bids_src_dir, overwrite)
+        bio.bids_init(bids_src_dir, overwrite)
 
     subject_dir_list = []
 
@@ -232,7 +233,7 @@ def main():
             if not first_pass:
 
                 # Get subject age and sex from representative DICOM header
-                dcm_info = bids_dcm_info(dcm_dir)
+                dcm_info = bio.bids_dcm_info(dcm_dir)
 
                 # Add line to participants TSV file
                 add_participant_record(bids_src_dir, sid, dcm_info['Age'], dcm_info['Sex'])
@@ -243,7 +244,7 @@ def main():
 
     if first_pass:
         # Create a template protocol dictionary
-        bids_create_prot_dict(prot_dict_json, prot_dict)
+        bio.bids_create_prot_dict(prot_dict_json, prot_dict)
 
     if not args.skip_if_pruning:
         print("Subject directories to prune:  " + ", ".join(subject_dir_list))
@@ -269,7 +270,9 @@ def bids_prune_intendedfors(bids_subj_dir, fmap_only):
         for name in files:
          
             # Only examine json files, ignore dataset_description, and only work in fmap directories if so specified
-            if os.path.splitext(name)[1] == ".json" and not name == "dataset_description.json" and (not fmap_only or os.path.basename(root) == "fmap"):
+            if (os.path.splitext(name)[1] == ".json" and
+                    not name == "dataset_description.json" and
+                    (not fmap_only or os.path.basename(root) == "fmap")):
 
                 with open(os.path.join(root, name), 'r+') as f:
 
@@ -327,7 +330,7 @@ def bids_run_conversion(conv_dir, first_pass, prot_dict, src_dir, sid, ses, clea
         # Note that glob returns arbitrarily ordered filenames, which must be sorted according to serNo
         # for the run ordering below to work correctly.
         filelisttmp = glob(os.path.join(conv_dir, '*.nii*'))
-        sernolist = [int(parse_dcm2niix_fname(file)['SerNo']) for file in filelisttmp]
+        sernolist = [int(bio.parse_dcm2niix_fname(file)['SerNo']) for file in filelisttmp]
         filelist = [file for _, file in sorted(zip(sernolist, filelisttmp))]
         
         # Infer run numbers accounting for duplicates.
@@ -338,7 +341,7 @@ def bids_run_conversion(conv_dir, first_pass, prot_dict, src_dir, sid, ses, clea
         for fc, src_nii_fname in enumerate(filelist):
 
             # Parse image filename into fields
-            info = parse_dcm2niix_fname(src_nii_fname)
+            info = bio.parse_dcm2niix_fname(src_nii_fname)
 
             # Check if we're creating new protocol dictionary
             if first_pass:
@@ -388,7 +391,7 @@ def bids_run_conversion(conv_dir, first_pass, prot_dict, src_dir, sid, ses, clea
 
                         # Create BIDS purpose directory
                         bids_purpose_dir = os.path.join(src_dir, bids_purpose)
-                        safe_mkdir(bids_purpose_dir)
+                        bio.safe_mkdir(bids_purpose_dir)
 
                         # Complete BIDS filenames for image and sidecar
                         if ses:
@@ -456,17 +459,17 @@ def bids_purpose_handling(bids_purpose, bids_intendedfor, seq_name,
     bids_bvec_fname = []
 
     # Load the JSON sidecar
-    info = bids_read_json(work_json_fname)
+    info = bio.bids_read_json(work_json_fname)
 
     if bids_purpose == 'func':
 
         if seq_name == 'EP':
 
             print('    EPI detected')
-            bids_events_template(bids_nii_fname, overwrite)
+            bio.bids_events_template(bids_nii_fname, overwrite)
 
             # Add taskname to BIDS JSON sidecar
-            bids_keys = parse_bids_fname(bids_nii_fname)
+            bids_keys = bio.parse_bids_fname(bids_nii_fname)
             if 'task' in bids_keys:
                 info['TaskName'] = bids_keys['task']
             else:
@@ -559,40 +562,16 @@ def bids_purpose_handling(bids_purpose, bids_intendedfor, seq_name,
     print('  Populating BIDS source directory')
 
     if bids_nii_fname:
-        safe_copy(work_nii_fname, str(bids_nii_fname), overwrite)
+        bio.safe_copy(work_nii_fname, str(bids_nii_fname), overwrite)
 
     if bids_json_fname:
-        bids_write_json(bids_json_fname, info, overwrite)
+        bio.bids_write_json(bids_json_fname, info, overwrite)
 
     if bids_bval_fname:
-        safe_copy(work_bval_fname, bids_bval_fname, overwrite)
+        bio.safe_copy(work_bval_fname, bids_bval_fname, overwrite)
 
     if bids_bvec_fname:
-        safe_copy(work_bvec_fname, bids_bvec_fname, overwrite)
-
-
-def bids_init(bids_src_dir, overwrite=False):
-    """
-    Initialize BIDS source directory
-
-    :param bids_src_dir: string
-        BIDS source directory
-    :param overwrite: string
-        Overwrite flag
-    :return True
-    """
-
-    # Create template JSON dataset description
-    datadesc_json = os.path.join(bids_src_dir, 'dataset_description.json')
-    meta_dict = dict({'BIDSVersion': "1.0.0",
-                      'License': "This data is made available under the Creative Commons BY-SA 4.0 International License.",
-                      'Name': "The dataset name goes here",
-                      'ReferencesAndLinks': "References and links for this dataset go here"})
-
-    # Write JSON file
-    bids_write_json(datadesc_json, meta_dict, overwrite)
-
-    return True
+        bio.safe_copy(work_bvec_fname, bids_bvec_fname, overwrite)
 
 
 def add_participant_record(studydir, subject, age, sex):
@@ -609,157 +588,19 @@ def add_participant_record(studydir, subject, age, sex):
     participants_tsv = os.path.join(studydir, 'participants.tsv')
     participant_id = 'sub-%s' % subject
 
-    if not create_file_if_missing(participants_tsv,'\t'.join(['participant_id', 'age', 'sex', 'group']) + '\n'):
-        # check if may be subject record already exists
+    if not bio.create_file_if_missing(participants_tsv, '\t'.join(['participant_id', 'age', 'sex', 'group']) + '\n'):
+
+        # Check if subject record already exists
         with open(participants_tsv) as f:
             f.readline()
             known_subjects = {l.split('\t')[0] for l in f.readlines()}
+
         if participant_id in known_subjects:
             return
 
     # Add a new participant
     with open(participants_tsv, 'a') as f:
         f.write('\t'.join(map(str, [participant_id, age.lstrip('0').rstrip('Y') if age else 'N/A', sex, 'control'])) + '\n')
-
-
-def create_file_if_missing(filename, content):
-    """
-    Create file if missing, so we do not override any possibly introduced changes
-
-    :param filename:
-    :param content:
-    :return:
-    """
-
-    if os.path.lexists(filename):
-        return False
-
-    dirname = os.path.dirname(filename)
-
-    if not os.path.exists(dirname):
-        os.makedirs(dirname)
-
-    with open(filename, 'w') as f:
-        f.write(content)
-
-    return True
-
-
-def bids_dcm_info(dcm_dir):
-    """
-    Extract relevant subject information from DICOM header
-    - Assumes only one subject present within dcm_dir
-
-    :param dcm_dir: directory containing all DICOM files or DICOM subfolders
-    :return dcm_info: DICOM header information dictionary
-    """
-
-    # Init the DICOM structure
-    ds = []
-
-    # Init the subject info dictionary
-    dcm_info = dict()
-
-    # Walk through dcm_dir looking for valid DICOM files
-    for subdir, dirs, files in os.walk(dcm_dir):
-        for file in files:
-
-            try:
-                ds = pydicom.read_file(os.path.join(subdir, file))
-            except IOError:
-                pass
-
-            # Break out if valid DICOM read
-            if ds:
-                break
-
-    if ds:
-
-        # Fill dictionary
-        # Note that DICOM anonymization tools sometimes clear these fields
-        if hasattr(ds, 'PatientSex'):
-            dcm_info['Sex'] = ds.PatientSex
-        else:
-            dcm_info['Sex'] = 'Unknown'
-
-        if hasattr(ds, 'PatientAge'):
-            dcm_info['Age'] = ds.PatientAge
-        else:
-            dcm_info['Age'] = 0
-
-    else:
-
-        print('* No DICOM header information found in %s' % dcm_dir)
-        print('* Confirm that DICOM images in this folder are uncompressed')
-        print('* Exiting')
-        sys.exit(1)
-
-    return dcm_info
-
-
-def parse_dcm2niix_fname(fname):
-    """
-    Parse dcm2niix filename into values
-    Filename format is '%n--%d--%q--%s' ie '<name>--<description>--<sequence>--<series #>'
-
-    :param fname: str
-        BIDS-style image or sidecar filename
-    :return info: dict
-    """
-
-    # Ignore containing directory and extension(s)
-    fname = strip_extensions(os.path.basename(fname))
-
-    # Create info dictionary
-    info = dict()
-
-    # Split filename at '--'s
-    vals = fname.split('--')
-
-    info['SubjName'] = vals[0]
-    info['SerDesc'] = vals[1]
-    info['SeqName'] = vals[2]
-
-    # Parse series string
-    # eg '10' or '10_e2' or '10_e2_ph'
-    ser_vals = vals[3].split('_')
-
-    info['SerNo'] = ser_vals[0]
-    if len(ser_vals) > 1:
-        info['EchoNo'] = ser_vals[1]
-    if len(ser_vals) > 2:
-        info['IsPhase'] = True
-
-    return info
-
-
-def parse_bids_fname(fname):
-    """
-    Parse BIDS filename into key-value pairs
-
-    :param fname:
-    :return:
-    """
-
-    # Init return dictionary
-    bids_keys = dict()
-
-    # Retain only basename without extensions (handle .nii.gz)
-    fname, _ = os.path.splitext(os.path.basename(fname))
-    fname, _ = os.path.splitext(fname)
-
-    kvs = fname.split('_')
-
-    for kv in kvs:
-
-        tmp = kv.split('-')
-
-        if len(tmp) > 1:
-            bids_keys[tmp[0]] = tmp[1]
-        else:
-            bids_keys['type'] = tmp[0]
-
-    return bids_keys
 
 
 def bids_add_run_number(bids_suffix, run_no):
@@ -794,72 +635,6 @@ def bids_add_run_number(bids_suffix, run_no):
     return new_bids_suffix
 
 
-def bids_events_template(bold_fname, overwrite=False):
-    """
-    Create a template events file for a corresponding BOLD imaging file
-    :param bold_fname: str
-        BOLD imaging filename (.nii.gz)
-    :param overwrite: bool
-        Overwrite flag
-    :return: Nothing
-    """
-    if "_bold.nii.gz" in bold_fname: #can have sbref.nii.gz here and you do not want overwrite it
-        events_fname = bold_fname.replace('_bold.nii.gz', '_events.tsv')
-        events_bname = os.path.basename(events_fname)
-
-        if os.path.isfile(events_fname):
-            if overwrite:
-                print('  Overwriting previous %s' % events_bname)
-                create_file = True
-            else:
-                print('  Preserving previous %s' % events_bname)
-                create_file = False
-        else:
-            print('  Creating %s' % events_fname)
-            create_file = True
-
-        if create_file:
-            fd = open(events_fname, 'w')
-            fd.write('onset\tduration\ttrial_type\tresponse_time\n')
-            fd.close()
-
-
-def strip_extensions(fname):
-    """
-    Remove one or more extensions from a filename
-    :param fname:
-    :return:
-    """
-
-    fstub, fext = os.path.splitext(fname)
-    if fext == '.gz':
-        fstub, fext = os.path.splitext(fstub)
-    return fstub
-
-
-def bids_load_prot_dict(prot_dict_json):
-    """
-    Read protocol translations from JSON file in DICOM directory
-
-    :param prot_dict_json: string
-        JSON protocol translation dictionary filename
-    :return:
-    """
-
-    if os.path.isfile(prot_dict_json):
-
-        # Read JSON protocol translator
-        json_fd = open(prot_dict_json, 'r')
-        prot_dict = json.load(json_fd)
-        json_fd.close()
-
-    else:
-
-        prot_dict = dict()
-
-    return prot_dict
-
-
 def bids_fmap_echotimes(src_phase_json_fname):
     """
     Extract TE1 and TE2 from mag and phase MEGE fieldmap pairs
@@ -869,34 +644,32 @@ def bids_fmap_echotimes(src_phase_json_fname):
     """
 
     # Init returned TEs
-    TE1, TE2 = 0.0, 0.0
+    te1, te2 = 0.0, 0.0
 
     if os.path.isfile(src_phase_json_fname):
 
         # Read phase image metadata
-        phase_dict = bids_read_json(src_phase_json_fname)
+        phase_dict = bio.bids_read_json(src_phase_json_fname)
 
         # Populate series info dictionary from dcm2niix output filename
-        info = parse_dcm2niix_fname(src_phase_json_fname)
+        info = bio.parse_dcm2niix_fname(src_phase_json_fname)
 
         # Magnitude 1 series number is one less than phasediff series number
         mag1_ser_no = str(int(info['SerNo']) - 1)
 
         # Construct dcm2niix mag1 JSON filename
         # Requires dicm2niix v1.0.20180404 or later for echo number suffix '_e1'
-        src_mag1_json_fname = info['SubjName'] + '--' +\
-                              info['SerDesc'] + '--' +\
-                              info['SeqName'] + '--' +\
-                              mag1_ser_no + '_e1.json'
+        src_mag1_json_fname = info['SubjName'] + '--' + info['SerDesc'] + '--' + \
+            info['SeqName'] + '--' + mag1_ser_no + '_e1.json'
         src_mag1_json_path = os.path.join(os.path.dirname(src_phase_json_fname), src_mag1_json_fname)
 
         # Read mag1 metadata
-        mag1_dict = bids_read_json(src_mag1_json_path)
+        mag1_dict = bio.bids_read_json(src_mag1_json_path)
 
-        # Add TE1 key and rename TE2 key
+        # Add te1 key and rename TE2 key
         if mag1_dict:
-            TE1 = mag1_dict['EchoTime']
-            TE2 = phase_dict['EchoTime']
+            te1 = mag1_dict['EchoTime']
+            te2 = phase_dict['EchoTime']
         else:
             print('*** Could not determine echo times multiecho fieldmap - using 0.0 ')
 
@@ -904,88 +677,7 @@ def bids_fmap_echotimes(src_phase_json_fname):
 
         print('* Fieldmap phase difference sidecar not found : ' + src_phase_json_fname)
 
-    return TE1, TE2
-
-
-def bids_create_prot_dict(prot_dict_json, prot_dict):
-    """
-    Write protocol translation dictionary template to JSON file
-    :param prot_dict_json: string
-        JSON filename
-    :param prot_dict: dictionary
-        Dictionary to write
-    :return:
-    """
-
-    if os.path.isfile(prot_dict_json):
-
-        print('* Protocol dictionary already exists : ' + prot_dict_json)
-        print('* Skipping creation of new dictionary')
-
-    else:
-
-        json_fd = open(prot_dict_json, 'w')
-        json.dump(prot_dict, json_fd, indent=4, separators=(',', ':'))
-        json_fd.close()
-
-        print('')
-        print('---')
-        print('New protocol dictionary created : %s' % prot_dict_json)
-        print('Remember to replace "EXCLUDE" values in dictionary with an appropriate image description')
-        print('For example "MP-RAGE T1w 3D structural" or "MB-EPI BOLD resting-state')
-        print('---')
-        print('')
-
-    return
-
-
-def bids_read_json(fname):
-    """
-    Safely read JSON sidecar file into a dictionary
-    :param fname: string
-        JSON filename
-    :return: dictionary structure
-    """
-
-    try:
-        fd = open(fname, 'r')
-        json_dict = json.load(fd)
-        fd.close()
-    except:
-        print('*** JSON sidecar not found - returning empty dictionary')
-        json_dict = dict()
-
-    return json_dict
-
-
-def bids_write_json(fname, meta_dict, overwrite=False):
-    """
-    Write a dictionary to a JSON file. Account for overwrite flag
-    :param fname: string
-        JSON filename
-    :param meta_dict: dictionary
-        Dictionary
-    :param overwrite: bool
-        Overwrite flag
-    :return:
-    """
-
-    bname = os.path.basename(fname)
-
-    if os.path.isfile(fname):
-        if overwrite:
-            print('    Overwriting previous %s' % bname)
-            create_file = True
-        else:
-            print('    Preserving previous %s' % bname)
-            create_file = False
-    else:
-        print('    Creating new %s' % bname)
-        create_file = True
-
-    if create_file:
-        with open(fname, 'w') as fd:
-            json.dump(meta_dict, fd, indent=4, separators=(',', ':'))
+    return te1, te2
 
 
 def bids_auto_run_no(file_list):
@@ -1000,7 +692,7 @@ def bids_auto_run_no(file_list):
     # Construct list of series descriptions and original numbers from file names
     ser_desc_list = []
     for fname in file_list:
-        info = parse_dcm2niix_fname(fname)
+        info = bio.parse_dcm2niix_fname(fname)
         ser_desc_list.append(info['SerDesc'])
 
     # Find unique ser_desc entries using sets
@@ -1040,43 +732,6 @@ def bids_build_intendedfor(sid, ses, bids_suffix):
         ifstr = os.path.join(bids_type, 'sub-' + sid + '_' + bids_name + '.nii.gz')
 
     return ifstr
-
-
-def safe_mkdir(dname):
-    """
-    Safely create a directory path
-    :param dname: string
-    :return:
-    """
-
-    if not os.path.isdir(dname):
-        os.makedirs(dname, exist_ok=True)
-
-
-def safe_copy(fname1, fname2, overwrite=False):
-    """
-    Copy file accounting for overwrite flag
-    :param fname1: str
-    :param fname2: str
-    :param overwrite: bool
-    :return:
-    """
-
-    bname1, bname2 = os.path.basename(fname1), os.path.basename(fname2)
-
-    if os.path.isfile(fname2):
-        if overwrite:
-            print('    Copying %s to %s (overwrite)' % (bname1, bname2))
-            create_file = True
-        else:
-            print('    Preserving previous %s' % bname2)
-            create_file = False
-    else:
-        print('    Copying %s to %s' % (bname1, bname2))
-        create_file = True
-
-    if create_file:
-        shutil.copy(fname1, fname2)
 
 
 def bids_add_intended_run(prot_dict, info, run_no):
