+++ bin/dcm2ndar
@@ -16,7 +16,7 @@ Mike Tyszka, Caltech Brain Imaging Center
 
 Dates
 ----
-2016-08-09 JMT Adapt from main.py
+2016-08-09 JMT Adapt from launcher.py
 
 MIT License
